pyproject.toml
[project]
name = "susde-arbitrage-monitor"
version = "0.1.0"
description = "SusDE arbitrage monitoring Telegram bot"
authors = [{name = "Your Name", email = "your.email@example.com"}]
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0.0",
    "playwright>=1.40.0",
    "web3>=7.0.0",
    "python-telegram-bot>=21.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "apscheduler>=3.11.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = []
